UI overhaul of the team details page * removed info page, making shifts the default * put team info above the tabnav * hide tabnav when the user is a regular volunteer and only show shifts
