+++ src/utils/path.test.ts
@@ -3,7 +3,6 @@ import {
   getEventShiftsPath,
   getQualificationDetailsPath,
   getQualificationsPath,
-  getTeamInfoPath,
   getTeamShiftsApiPath,
   getTeamShiftsPath,
   getTeamVolunteersPath,
@@ -17,21 +16,12 @@ import {
   getCreateTeamPath
 } from './path';
 
-describe('getTeamInfoPath', () => {
-  it('should return the correct path for team info', () => {
-    const eventSlug = 'event-123';
-    const teamSlug = 'team-abc';
-    const result = getTeamInfoPath(eventSlug, teamSlug);
-    expect(result).toBe('/event/event-123/team/team-abc');
-  });
-});
-
 describe('getTeamShiftsPath', () => {
   it('should return the correct path for team shifts', () => {
     const eventSlug = 'event-123';
     const teamSlug = 'team-abc';
     const result = getTeamShiftsPath(eventSlug, teamSlug);
-    expect(result).toBe('/event/event-123/team/team-abc/shifts');
+    expect(result).toBe('/event/event-123/team/team-abc');
   });
 });
